Improve codeclimate satisfaction
