+++ app/helpers/results_parser_helper.rb
@@ -1,0 +1,34 @@
+module ResultsParser
+  def count_totals(items)
+    response  = { hits: 0, conversions: 0, upsells: 0, sales: 0, total_upsells: 0, total_sales: 0, avg_order_value: 0, hits_to_orders: 0 }
+
+    items.each { |source, values| values.each { |key, value| response[key] += value } }
+
+    total_orders  = response[:conversions] + response[:upsells] + 0.0
+    value_orders  = response[:total_sales] + 0.0
+
+    response[:avg_order_value]  = total_orders == 0     ? 0 : value_orders / total_orders
+    response[:hits_to_orders]   = response[:hits] == 0  ? 0 : total_orders / response[:hits]
+
+    response
+  end
+
+  def compose_view_hash(items)
+    response  = {}
+
+    items.each do |source, values|
+      response[source]  = {
+        hits:             values[:hits],
+        conversions:      values[:sales],
+        upsells:          values[:upsells],
+        total_upsells:    values[:total_upsells],
+        sales:            values[:total_sales],
+        total_sales:      values[:total_sales] + values[:total_upsells],
+        hits_to_orders:   values[:conversions],
+        avg_order_value:  values[:avg_order_value],
+      }
+    end
+
+    response
+  end
+end
