+++ bootstrap.rb
@@ -10,5 +10,8 @@ set :database_file, "#{APPLICATION_PATH}/config/database.yml"
 # Require app/models
 Dir["#{APPLICATION_PATH}/app/models/*.rb"].each { |file| require file }
 
+# Require app/helpers
+Dir["#{APPLICATION_PATH}/app/helpers/*_helper.rb"].each { |file| require file }
+
 # Require libs
 Dir["#{APPLICATION_PATH}/lib/*.rb"].each { |file| require file }
